pyproject.toml
[tool.poetry]
name = "flibusta-telegram-bot"
version = "0.1.0"
description = "telegram bot for download books from flibusta"
authors = ["Evgeniy Korzunov <korzunov.e.v@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = "^3.12"
python-telegram-bot = {version = "^21.9", extras = ["job-queue"]}
python-dotenv = "^1.0.1"
requests = "^2.31"
beautifulsoup4 = "^4.12"
lxml = "^4.9"
json-log-formatter = "^1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
